requirements.txt
Django==1.5.5
Mezzanine==3.0.4
MySQL-python==1.2.4
Pillow==2.2.2
WebOb==1.3.1
WebTest==2.0.11
argparse==1.2.1
beautifulsoup4==4.3.2
bleach==1.2.2
chardet==2.2.1
django-appconf==0.6
django-extensions==1.3.2
django-haystack==2.1.0
django-newsletter==0.5.1
django-staticfiles==1.2.1
django-tinymce==1.5.2
django-webtest==1.7.5
filebrowser-safe==0.3.1
future==0.9.0
grappelli-safe==0.3.3
html5lib==0.95
oauthlib==0.6.0
psycopg2==2.5.2
pyelasticsearch==0.6.1
python-dateutil==2.2
pytz==2013.8
requests==1.2.3
requests-oauthlib==0.3.3
simplejson==3.3.2
six==1.5.2
sorl-thumbnail==11.12
surlex==0.2.0
vobject==0.6.6
waitress==0.8.8
wsgiref==0.1.2
